fix: updated filename for drive-v3
